Texture Manager:  Loads each texture once a material.
Material now contains the information of the texture_id, and hence a simple call
to glBindTexture with the materials texture_id, will suffice to use the texture.
TODO : When multiple textures are used, texture_manager willinform the shader of
the texture units to use.  The set_shader_uniforms will update the shader to use
the right textures (normal_map, ambient_map, diffuse_map, etc ...)